requirements.txt
fastapi
uvicorn[standard]
python-multipart
pytest
httpx
fastapi-cache2[redis]